Use SceneStorage instead of CoreData to save location info.
Required subclassing CLLocation to make this work because @SceneStorage can only store RawRepresentable types.
CLLocation is not RawRepresentable, but it is codable, so I was able to use NSKeyedArchiver to implement the
RawRepresentable protocol on my subclass.

Because RawRepresentable requires implementing an initializer, it was not possible to simply extend CLLocation with the protocol.